Extracts hooks to access atoms & selectors and reduce duplication elsewhere
